requirements.txt
functions-framework==3.*
python-telegram-bot==21.7
deepgram-sdk==3.4.0